requirements.txt
Django==5.1
djangorestframework
djangorestframework-simplejwt
google-auth
python-dotenv
requests